backend/requirements.txt
Django==4.2.15
django-cors-headers==4.4.0
django-taggit==6.0.0
djangorestframework==3.15.2
pillow==10.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
